+++ src/app/(user)/(service)/my-page/liked/page.tsx
@@ -1,7 +1,7 @@
 "use client";
 
 import FeedbackList from "@/src/components/feedback/FeedbackList";
-import LikedFeedbackList from "@/src/components/my-page/LikedFeedbackList";
+import LikedFeedbackList from "@/src/features/my-page/my-page-feedback/components/liked-feedback-list";
 import MyPlacePagination from "@/src/features/my-place/my-place-common/components/my-place-pagination";
 import { ArrowLeftIcon, ArrowRightIcon } from "@/src/components/utils/icons";
 import client from "@/src/lib/api/client";
