+++ src/features/my-page/my-page-direction/components/recent-destinations-item.tsx
@@ -1,12 +1,12 @@
 "use client";
 
 import { useSession } from "next-auth/react";
-import { StarIcon } from "../utils/icons";
 import { useMutation, useQuery, useQueryClient } from "@tanstack/react-query";
 import client from "@/src/lib/api/client";
-import { Destination } from "./RecentDestinations";
 import { IFeedbackListItem } from "@/src/types/feedback/feedbackList";
 import { useRouter } from "next/navigation";
+import { Destination } from "./recent-destinations";
+import { StarIcon } from "lucide-react";
 
 export interface DestinationItem {
   id?: string;
